mongoConnection.js moved to UTILS and references to this file changed
